webservice-maintenance-action/requirements.txt
confluent-kafka 
numpy
pandas
fastapi 
uvicorn 
pydantic 
requests
opensearch
psycopg2-binary
geopy
